requirements.txt
streamlit
python-dotenv
google-generativeai
Pillow
streamlit-lottie
requests
